requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
fastapi==0.115.0
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
gotrue==2.9.2
greenlet==3.1.1
gritql==0.1.5
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
hyperframe==6.0.1
idna==3.10
Jinja2==3.1.4
jiter==0.6.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.3
langchain-cli==0.0.31
langchain-community==0.3.2
langchain-core==0.3.10
langchain-text-splitters==0.3.0
langserve==0.3.0
langsmith==0.1.132
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.1
marshmallow==3.22.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4
numpy==1.26.4
openai==1.51.1
orjson==3.10.7
packaging==24.1
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
postgrest==0.17.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
PyYAML==6.0.2
realtime==2.0.5
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.0
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
sse-starlette==1.8.2
starlette==0.38.6
storage3==0.8.1
StrEnum==0.4.15
supabase==2.9.0
supafunc==0.6.1
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.1
tomlkit==0.12.5
torch==2.4.1
tqdm==4.66.5
transformers==4.45.2
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.23.2
websockets==13.1
yarl==1.13.1
